test: move gitattributes policy check
