+++ tests/docs_workspace_membership.rs
@@ -12,27 +12,6 @@ fn claude_and_agents_docs_are_byte_identical() {
 }
 
 #[test]
-fn gitattributes_preserves_repo_line_ending_policy() {
-    let root = env!("CARGO_MANIFEST_DIR");
-    let attrs = fs::read_to_string(format!("{root}/.gitattributes"))
-        .expect("Failed to read .gitattributes");
-
-    let required = [
-        "* text=auto eol=lf",
-        "*.bat text eol=crlf",
-        "*.cmd text eol=crlf",
-        "*.ps1 text eol=crlf",
-    ];
-
-    for pattern in required {
-        assert!(
-            attrs.lines().any(|line| line.trim() == pattern),
-            ".gitattributes must contain `{pattern}`"
-        );
-    }
-}
-
-#[test]
 fn architecture_docs_list_all_workspace_crates() {
     let root = env!("CARGO_MANIFEST_DIR");
